Rename internal area-related history variables for clarity.
FATES_PATCHAREA_LU and FATES_PATCHAREA_AP are not in units of m2, but rather m2/m2. It is thus misleading for them to be named "AREA". In the interest of not messing up user workflows, I'm not changing those---at least for now. However, this commit renames related internal variables in the history code for clarity.

Specifically, "area" is replaced with "fracarea" (fractional area, relative to site area) in the following:
- ih_area_si_landuse
- ih_area_si_age
- hio_area_si_landuse
- hio_fracarea_si_age